# openapi-current.yaml
openapi: 3.0.0
info:
  title: Donation Matching Portal API (Seva Sahayog Foundation)
  version: 1.0.0
  description: |
    REST API for the Seva Sahayog donation matching portal.

    Roles: DONOR, RECEIVER, ADMIN.
    All endpoints except /api/auth/* require a JWT bearer token.
    Authorization is enforced server-side via @PreAuthorize and ownership checks.
servers:
  - url: /
...
paths:
...
  /api/requirements:
    post:
      tags: [Requirements]
      summary: Create a requirement
      description: RECEIVER only. Creates requirement with status SUBMITTED.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementRequest'
      responses:
        '201':
          description: Requirement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Requirements]
      summary: Search approved requirements (discovery)
      description: |
        DONOR sees only APPROVED requirements and may NOT pass a status filter (400).
        ADMIN sees all statuses and may filter by status.
        RECEIVER is forbidden.
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Category'
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: query
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          required: false
          description: ADMIN only. Rejected with 400 for non-admin roles.
          schema:
            $ref: '#/components/schemas/RequirementStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Whitelisted to createdAt.
          schema:
            type: string
            default: createdAt,desc
      responses:
        '200':
          description: Paged results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/requirements/my:
    get:
      tags: [Requirements]
      summary: List the logged-in receiver's own requirements
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Paged own requirements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/requirements/{id}:
    get:
      tags: [Requirements]
      summary: View a single requirement
      description: RECEIVER or ADMIN. RECEIVER must own the requirement (403 otherwise).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Requirement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not found or not owned by caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Requirements]
      summary: Update an own requirement (partial)
      description: RECEIVER only, owner required. Returns 204 with no body.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementUpdateRequest'
      responses:
        '204':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not found or not owned by caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/requirements/{id}/status:
    patch:
      tags: [Requirements]
      summary: Admin approve/reject a requirement
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementStatusUpdateRequest'
      responses:
        '200':
          description: Updated requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing, malformed, expired or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Role not permitted or resource not owned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation failed or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Category:
      type: string
      enum: [FOOD, CLOTHING, EDUCATION, MEDICAL, FURNITURE, ELECTRONICS, HOUSEHOLD, HYGIENE, OTHER]
    QuantityUnit:
      type: string
      enum: [PIECES, KG, LITRES, BAGS, BOXES, PACKETS, SETS]
...
    Urgency:
      type: string
      enum: [LOW, MEDIUM, HIGH]
...
    RequirementStatus:
      type: string
      enum: [SUBMITTED, APPROVED, FULFILLED, REJECTED, CANCELLED]
...
    RequirementRequest:
      type: object
      required: [title, description, category, quantity, quantityUnit, city, urgency]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        category:
          $ref: '#/components/schemas/Category'
        quantity:
          type: number
          format: double
          exclusiveMinimum: 0
          description: At most 9 integer digits and 3 decimal places.
        quantityUnit:
          $ref: '#/components/schemas/QuantityUnit'
        city:
          type: string
          maxLength: 100
        locality:
          type: string
          maxLength: 100
        pincode:
          type: string
          maxLength: 20
        urgency:
          $ref: '#/components/schemas/Urgency'
    RequirementUpdateRequest:
      description: Partial update; all fields optional. Server-controlled fields are ignored.
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        category:
          $ref: '#/components/schemas/Category'
        quantity:
          type: number
          format: double
          exclusiveMinimum: 0
        quantityUnit:
          $ref: '#/components/schemas/QuantityUnit'
        city:
          type: string
          maxLength: 100
        locality:
          type: string
          maxLength: 100
        pincode:
          type: string
          maxLength: 20
        urgency:
          $ref: '#/components/schemas/Urgency'
    RequirementStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/RequirementStatus'

    # ---------- Responses ----------
...
    UserSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    RequirementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        quantity:
          type: number
          format: double
        quantityUnit:
          $ref: '#/components/schemas/QuantityUnit'
        city:
          type: string
        locality:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
        urgency:
          $ref: '#/components/schemas/Urgency'
        status:
          $ref: '#/components/schemas/RequirementStatus'
        receiver:
          $ref: '#/components/schemas/UserSummaryResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # ---------- Pagination (Spring Data Page JSON shape) ----------
    Page:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageable:
          type: object
          properties:
            pageNumber:
              type: integer
              format: int32
            pageSize:
              type: integer
              format: int32
            sort:
              $ref: '#/components/schemas/Sort'
            offset:
              type: integer
              format: int64
            paged:
              type: boolean
            unpaged:
              type: boolean
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...
    RequirementPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RequirementResponse'
...